requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
django-cors-headers==3.4.0
django==3.1
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
drf-yasg2==1.19.4
pillow==7.2.0
psycopg2==2.8.6
uwsgi==2.0.19.1
